+++ src/main/java/org/sopt/solply_server/domain/course/entity/Course.java
@@ -15,7 +15,6 @@ import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-@EntityListeners(AuditingEntityListener.class)
 @Table(name = "courses",
         indexes = {
                 @Index(name = "idx_course_town_id", columnList = "town_id"),
